re-structured sub-modules
re-structured sub-modules & updated docs
